renamed libraries to typescript folder
