+++ src/api/core/_credentialsFactory.ts
@@ -1,7 +1,7 @@
 import { AxiosResponse, default as axios } from 'axios';
 import { access, constants, readFileSync, writeFileSync } from 'fs';
 import { dirname } from 'path';
-import { AcountNumberString, Credentials, IAccount, IAccounts, ITime, QuestradeAPIOptions } from '../../libraries';
+import { AcountNumberString, Credentials, IAccount, IAccounts, ITime, QuestradeAPIOptions } from '../../typescript';
 import { sync } from '../../utils/mkdirp';
 import { _axiosApiGet } from './_axiosApi';
