+++ src/typescript/IActivities.ts
