+++ src/typescript/ICandles.ts
