+++ src/typescript/ICreds.ts
