+++ src/typescript/IEquitySymbols.ts
