+++ src/typescript/IQuotes.ts
