+++ src/typescript/IStockSymbol.ts
