+++ src/typescript/Methode.ts
