+++ src/typescript/Optionals.ts
@@ -1,9 +1,5 @@
-import {
-  HistoricalDataGranularity,
-  OptionType,
-  OrderStateFilterType,
-} from 'questrade-api-enumerations';
-import { DateTime } from '../libraries';
+import { HistoricalDataGranularity, OptionType, OrderStateFilterType } from 'questrade-api-enumerations';
+import { DateTime } from '.';
 
 // - Copyright (c) Benjamin Vincent Kasapoglu (Luxcium). All rights reserved.
 // - Licensed under the MIT License.
