+++ src/typescript/void_0.ts
